+++ library/src/regex.ts
@@ -1,4 +1,9 @@
 /**
+ * [BIC] (https://en.wikipedia.org/wiki/ISO_9362) regex.
+ */
+export const BIC_REGEX = /^[A-Z]{6}(?!00)[A-Z\d]{2}(?:[A-Z\d]{3})?$/u;
+
+/**
  * [Cuid2](https://github.com/paralleldrive/cuid2) regex.
  */
 export const CUID2_REGEX = /^[a-z][\da-z]*$/u;
@@ -74,8 +79,3 @@ export const ULID_REGEX = /^[\da-hjkmnp-tv-z]{26}$/iu;
  * [UUID](https://en.wikipedia.org/wiki/Universally_unique_identifier) regex.
  */
 export const UUID_REGEX = /^[\da-f]{8}(?:-[\da-f]{4}){3}-[\da-f]{12}$/iu;
-
-/**
- * [BIC] (https://en.wikipedia.org/wiki/ISO_9362) regex.
- */
-export const BIC_REGEX = /^[A-Z]{6}(?!00)[A-Z\d]{2}(?:[A-Z\d]{3})?$/u;